Split Assets screen into components
